requirements.txt
torch==2.2.2
transformers==4.40.2
pyyaml==6.0.2
pydantic==2.9.2
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
accelerate==0.30.1
sentencepiece==0.2.0
pytest==8.3.3